+++ .env.production
@@ -1,5 +1,5 @@
-DB_USER=
-DB_HOST=
-DB_NAME=
-DB_PASSWORD=
-DB_PORT=
+DB_USER=my_blog_user
+DB_HOST=localhost
+DB_NAME=my_blog
+DB_PASSWORD=my_blog_user_password
+DB_PORT=8001
