+++ components/core/date.tsx
@@ -1,6 +1,10 @@
 import { parseISO, format } from 'date-fns';
 
-export default function Date({ dateString }) {
+interface DateOps {
+    dateString: string;
+}
+
+export default function Date({ dateString }: DateOps) {
     const date = parseISO(dateString);
     return <time dateTime={dateString}>{format(date, 'LLLL d, yyyy')}</time>;
 }
