+++ components/layout.tsx
@@ -8,7 +8,12 @@ import style from './layout.module.scss'
 export const BLOG_NAME = 'My Dragon is Angry'
 export const AVATAR_SRC = "/images/alencia_epic7.jpg"
 
-export default function Layout({ children, home }) {
+interface LayoutProps {
+    children: any;
+    home?: any;
+}
+
+export default function Layout({ children, home = false }: LayoutProps) {
     return (
         <div className={style.container} >
             <Head>
