+++ pages/_app.tsx
@@ -1,0 +1,8 @@
+import { AppProps } from 'next/dist/next-server/lib/router/router'
+import '../styles/globals.css'
+
+function MyApp({ Component, pageProps }: AppProps) {
+  return <Component {...pageProps} />
+}
+
+export default MyApp
